+++ lshash/point.h
@@ -11,10 +11,10 @@ struct Point {
 	double d[DIMS];
 
 	u64 operator*(const Point &_p) {
-		u64 res = 0;
+		double res = 0;
 		for(u_int i = 0; i < DIMS; ++i)
 			res += _p.d[i] * d[i];
-		return res;
+		return (u64)res;
 	}
 };
